2005-10-24 02:54 UTC+0200 Przemyslaw Czerpak (druzus/at/priv.onet.pl)   * xharbour/contrib/libnf/Makefile     + added missing fttext.c   * xharbour/contrib/libnf/fttext.c     * GCC warning cleaned
  * xharbour/contrib/xwt/Rules.make
  * xharbour/contrib/xwt/src/xwt_gtk/xwt_calendar.c
  * xharbour/contrib/xwt/src/xwt_gtk/xwt_gtk.c
  * xharbour/contrib/xwt/src/xwt_gtk/xwt_gtk_browse.c
  * xharbour/contrib/xwt/src/xwt_gtk/xwt_gtk_calendar.c
  * xharbour/contrib/xwt/src/xwt_gtk/xwt_gtk_checkbox.c
  * xharbour/contrib/xwt/src/xwt_gtk/xwt_gtk_colorselect.c
  * xharbour/contrib/xwt/src/xwt_gtk/xwt_gtk_combobox.c
  * xharbour/contrib/xwt/src/xwt_gtk/xwt_gtk_fileselect.c
  * xharbour/contrib/xwt/src/xwt_gtk/xwt_gtk_fontselect.c
  * xharbour/contrib/xwt/src/xwt_gtk/xwt_gtk_framewnd.c
  * xharbour/contrib/xwt/src/xwt_gtk/xwt_gtk_image.c
  * xharbour/contrib/xwt/src/xwt_gtk/xwt_gtk_listbox.c
  * xharbour/contrib/xwt/src/xwt_gtk/xwt_gtk_progressbar.c
  * xharbour/contrib/xwt/src/xwt_gtk/xwt_gtk_radiobutton.c
  * xharbour/contrib/xwt/src/xwt_gtk/xwt_gtk_textbox.c
  * xharbour/contrib/xwt/src/xwt_gtk/xwt_gtk_togglebutton.c
  * xharbour/contrib/xwt/src/xwt_gtk/xwt_gtk_treelist.c
  * xharbour/contrib/xwt/src/xwt_gtk/xwt_gtk_viewport.c
    * updated included headers

  * xharbour/include/Makefile
    + added hbvmopt.h

+++ xharbour/contrib/libnf/fttext.c
@@ -1,5 +1,5 @@
 /*
- * $Id: fttext.c,v 1.8 2005/05/16 21:45:39 andijahja Exp $
+ * $Id: fttext.c,v 1.9 2005/10/24 02:14:11 ronpinkas Exp $
  */
 
 /*
@@ -55,12 +55,12 @@
 
 #include "hbapi.h"
 #include "hbapifs.h"
+#include "hbapiitm.h"
 #include "hbfast.h"
+#include "hbstack.h"
 #include "hbvmpub.h"
 #include "hbpcode.h"
 #include "hbinit.h"
-#include "hbapiitm.h"
-#include "hbstack.h"
 
 #define __PRG_SOURCE__ __FILE__
 #undef HB_PRG_PCODE_VER
@@ -454,7 +454,7 @@ HB_FUNC ( FT_FACTIVE )
 //------------------------------------------------------------------------------
 HB_FUNC ( FT_FBUFFERSIZE )
 {
-   LONG uNewBuff = ISNUM(1) ? hb_parnl(1) : uBuffSize;
+   LONG uNewBuff = ISNUM(1) ? hb_parnl(1) : ( LONG ) uBuffSize;
 
    hb_retnl( uBuffSize );
